wiring in of cobertura test coverage framework, additional work on FileManagement package
