+++ src/main/java/FileSieve/BusinessLogic/FileManagement/FileManager.java
@@ -1,14 +1,13 @@
 package FileSieve.BusinessLogic.FileManagement;
 
+import java.beans.PropertyChangeListener;
 import java.io.IOException;
 import java.nio.file.Path;
 
 /**
  * Interface defining file management operations
- *
- * @param <T>   Type returned by the interface's copyProvider method
  */
-public interface FileManager<T> {
+public interface FileManager {
 
     /**
      * Method for copying a file or creating a folder to/within a target pathname. Method may handle the copy operation
@@ -19,10 +18,9 @@ public interface FileManager<T> {
      * @param recursionEnabled  recursive search for files within subfolders
      * @param overwriteExisting indicates if existing files in the target path should be overwritten
      * @param ifSizeDiffers     if "overwriteExisting" argument is true, overwrites existing files only if their size differs
-     * @return                  implementer-defined object, for indicating copy success or returning a worker object
-     *                          capable of handling a longer-running copy operation
+     * @return                  boolean true if the copy operation is begun or has succeeded, false if not
      */
-    public T copyProvider(Path sourcePathname, Path targetPathname, boolean recursionEnabled, boolean overwriteExisting, boolean ifSizeDiffers);
+    public boolean copyPathname(Path sourcePathname, Path targetPathname, boolean recursionEnabled, boolean overwriteExisting, boolean ifSizeDiffers);
 
     /**
      * Method for deleting a file or folder
